Refactor effect (manager) stuff
