+++ include/effectinitializers.h
@@ -1,0 +1,263 @@
+//+--------------------------------------------------------------------------
+//
+// File:        effectinitializers.h
+//
+// NightDriverStrip - (c) 2023 Plummer's Software LLC.  All Rights Reserved.
+//
+// This file is part of the NightDriver software project.
+//
+//    NightDriver is free software: you can redistribute it and/or modify
+//    it under the terms of the GNU General Public License as published by
+//    the Free Software Foundation, either version 3 of the License, or
+//    (at your option) any later version.
+//
+//    NightDriver is distributed in the hope that it will be useful,
+//    but WITHOUT ANY WARRANTY; without even the implied warranty of
+//    MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+//    GNU General Public License for more details.
+//
+//    You should have received a copy of the GNU General Public License
+//    along with Nightdriver.  It is normally found in copying.txt
+//    If not, see <https://www.gnu.org/licenses/>.
+//
+//
+// Description:
+//
+//    Declarations of different types for the effect initializers included
+//    in effectinitializers.cpp.
+//
+// History:     Apr-13-2019         Davepl      Created for NightDriverStrip
+//
+//---------------------------------------------------------------------------
+
+#pragma once
+
+#include "globals.h"
+
+// Palettes used by a number of effects
+
+const CRGBPalette16 BlueColors_p =
+{
+    CRGB::DarkBlue,
+    CRGB::MediumBlue,
+    CRGB::Blue,
+    CRGB::MediumBlue,
+    CRGB::DarkBlue,
+    CRGB::MediumBlue,
+    CRGB::Blue,
+    CRGB::MediumBlue,
+    CRGB::DarkBlue,
+    CRGB::MediumBlue,
+    CRGB::Blue,
+    CRGB::MediumBlue,
+    CRGB::DarkBlue,
+    CRGB::MediumBlue,
+    CRGB::Blue,
+    CRGB::MediumBlue
+};
+
+const CRGBPalette16 RedColors_p =
+{
+    CRGB::Red,
+    CRGB::DarkRed,
+    CRGB::DarkRed,
+    CRGB::DarkRed,
+
+    CRGB::Red,
+    CRGB::DarkRed,
+    CRGB::DarkRed,
+    CRGB::DarkRed,
+
+    CRGB::Red,
+    CRGB::DarkRed,
+    CRGB::DarkRed,
+    CRGB::DarkRed,
+
+    CRGB::Red,
+    CRGB::DarkRed,
+    CRGB::DarkRed,
+    CRGB::OrangeRed
+};
+
+const CRGBPalette16 GreenColors_p =
+{
+    CRGB::Green,
+    CRGB::DarkGreen,
+    CRGB::DarkGreen,
+    CRGB::DarkGreen,
+
+    CRGB::Green,
+    CRGB::DarkGreen,
+    CRGB::DarkGreen,
+    CRGB::DarkGreen,
+
+    CRGB::Green,
+    CRGB::DarkGreen,
+    CRGB::DarkGreen,
+    CRGB::DarkGreen,
+
+    CRGB::Green,
+    CRGB::DarkGreen,
+    CRGB::DarkGreen,
+    CRGB::LimeGreen
+};
+
+const CRGBPalette16 PurpleColors_p =
+{
+    CRGB::Purple,
+    CRGB::Maroon,
+    CRGB::Violet,
+    CRGB::DarkViolet,
+
+    CRGB::Purple,
+    CRGB::Maroon,
+    CRGB::Violet,
+    CRGB::DarkViolet,
+
+    CRGB::Purple,
+    CRGB::Maroon,
+    CRGB::Violet,
+    CRGB::DarkViolet,
+
+    CRGB::Pink,
+    CRGB::Maroon,
+    CRGB::Violet,
+    CRGB::DarkViolet,
+};
+
+const CRGBPalette16 RGBColors_p =
+{
+    CRGB::Red,
+    CRGB::Green,
+    CRGB::Blue,
+    CRGB::Red,
+    CRGB::Green,
+    CRGB::Blue,
+    CRGB::Red,
+    CRGB::Green,
+    CRGB::Blue,
+    CRGB::Red,
+    CRGB::Green,
+    CRGB::Blue,
+    CRGB::Red,
+    CRGB::Green,
+    CRGB::Blue,
+    CRGB::Blue
+};
+
+const CRGBPalette16 MagentaColors_p =
+{
+    CRGB::Pink,
+    CRGB::DeepPink,
+    CRGB::HotPink,
+    CRGB::LightPink,
+    CRGB::LightCoral,
+    CRGB::Purple,
+    CRGB::MediumPurple,
+    CRGB::Magenta,
+    CRGB::DarkMagenta,
+    CRGB::DarkSalmon,
+    CRGB::MediumVioletRed,
+    CRGB::Pink,
+    CRGB::DeepPink,
+    CRGB::HotPink,
+    CRGB::LightPink,
+    CRGB::Magenta};
+
+const CRGBPalette16 spectrumBasicColors =
+{
+    CRGB(0xFD0E35), // Red
+    CRGB(0xFF8833), // Orange
+    CRGB(0xFFEB00), // Middle Yellow
+    CRGB(0xAFE313), // Inchworm
+    CRGB(0x3AA655), // Green
+    CRGB(0x8DD9CC), // Middle Blue Green
+    CRGB(0x0066FF), // Blue III
+    CRGB(0xDB91EF), // Lilac
+    CRGB(0xFD0E35), // Red
+    CRGB(0xFF8833), // Orange
+    CRGB(0xFFEB00), // Middle Yellow
+    CRGB(0xAFE313), // Inchworm
+    CRGB(0x3AA655), // Green
+    CRGB(0x8DD9CC), // Middle Blue Green
+    CRGB(0x0066FF), // Blue III
+    CRGB(0xDB91EF)  // Lilac
+};
+
+const CRGBPalette16 spectrumAltColors =
+{
+    CRGB::Red,
+    CRGB::OrangeRed,
+    CRGB::Orange,
+    CRGB::Green,
+    CRGB::ForestGreen,
+    CRGB::Cyan,
+    CRGB::Blue,
+    CRGB::Indigo,
+    CRGB::Red,
+    CRGB::OrangeRed,
+    CRGB::Orange,
+    CRGB::Green,
+    CRGB::ForestGreen,
+    CRGB::Cyan,
+    CRGB::Blue,
+    CRGB::Indigo,
+};
+
+const CRGBPalette16 USAColors_p =
+{
+    CRGB::Blue,
+    CRGB::Blue,
+    CRGB::Blue,
+    CRGB::Blue,
+    CRGB::Blue,
+    CRGB::Red,
+    CRGB::White,
+    CRGB::Red,
+    CRGB::White,
+    CRGB::Red,
+    CRGB::White,
+    CRGB::Red,
+    CRGB::White,
+    CRGB::Red,
+    CRGB::White,
+    CRGB::Red,
+};
+
+const CRGBPalette16 rainbowPalette(RainbowColors_p);
+
+// Adds a default and JSON effect factory for a specific effect number and type.
+//   All parameters beyond effectNumber and effectType will be passed on to the default effect constructor.
+#define ADD_EFFECT(effectNumber, effectType, ...) \
+    g_ptrEffectFactories->AddEffect(effectNumber, \
+        []()                                 ->std::shared_ptr<LEDStripEffect> { return make_shared_psram<effectType>(__VA_ARGS__); }, \
+        [](const JsonObjectConst& jsonObject)->std::shared_ptr<LEDStripEffect> { return make_shared_psram<effectType>(jsonObject); }\
+    )
+
+// Adds a default and JSON effect factory for a specific effect number/type.
+//   All parameters beyond effectNumber and effectType will be passed on to the default effect constructor.
+//   The default effect will be disabled upon creation, so will not show until enabled.
+#define ADD_EFFECT_DISABLED(effectNumber, effectType, ...) \
+    ADD_EFFECT(effectNumber, effectType, __VA_ARGS__).LoadDisabled = true
+
+// Used by ADD_STARRY_NIGHT_EFFECT
+std::shared_ptr<LEDStripEffect> CreateStarryNightEffectFromJSON(const JsonObjectConst& jsonObject);
+
+// Adds a default and JSON effect factory for a StarryNightEffect with a specific star type.
+//   All parameters beyond starType will be passed on to the default StarryNightEffect constructor for the indicated star type.
+#define ADD_STARRY_NIGHT_EFFECT(starType, ...) \
+    g_ptrEffectFactories->AddEffect(EFFECT_STRIP_STARRY_NIGHT, \
+        []()                                 ->std::shared_ptr<LEDStripEffect> { return make_shared_psram<StarryNightEffect<starType>>(__VA_ARGS__); }, \
+        [](const JsonObjectConst& jsonObject)->std::shared_ptr<LEDStripEffect> { return CreateStarryNightEffectFromJSON(jsonObject); }\
+    )
+
+// Adds a default and JSON effect factory for a StarryNightEffect with a specific star type.
+//   All parameters beyond starType will be passed on to the default StarryNightEffect constructor for the indicated star type.
+//   The default effect will be disabled upon creation, so will not show until enabled.
+#define ADD_STARRY_NIGHT_EFFECT_DISABLED(starType, ...) \
+    ADD_STARRY_NIGHT_EFFECT(starType, __VA_ARGS__).LoadDisabled = true
+
+// Defines used by some StarryNightEffect instances
+
+#define STARRYNIGHT_PROBABILITY 1.0
+#define STARRYNIGHT_MUSICFACTOR 1.0
